Move language specific UIs to subfolders
