+++ packages/ui/src/lang/cairo/AccountControls.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-  import HelpTooltip from '../HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
   import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo';
   import { account, Account, infoDefaults } from '@openzeppelin/wizard-cairo';
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import InfoSection from './InfoSection.svelte';
-  import { error } from '../error-tooltip';
+  import { error } from '../../error-tooltip';
 
   export const opts: Required<KindedOptions['Account']> = {
     kind: 'Account',
