+++ packages/ui/src/lang/cairo/ERC1155Controls.svelte
@@ -1,11 +1,12 @@
 <script lang="ts">
-  import HelpTooltip from './HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
-  import type { KindedOptions } from '@openzeppelin/wizard';
-  import { erc1155, infoDefaults } from '@openzeppelin/wizard';
+  import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo';
+  import { erc1155, infoDefaults } from '@openzeppelin/wizard-cairo';
 
   import AccessControlSection from './AccessControlSection.svelte';
-  import UpgradeabilitySection from './UpgradeabilitySection.svelte';
+  import UpgradeabilityField from './UpgradeabilityField.svelte';
+  import RoyaltyInfoSection from './RoyaltyInfoSection.svelte';
   import InfoSection from './InfoSection.svelte';
 
   export let opts: Required<KindedOptions['ERC1155']> = {
@@ -13,6 +14,8 @@
     ...erc1155.defaults,
     info: { ...infoDefaults }, // create new object since Info is nested
   };
+
+  export let errors: undefined | OptionsErrorMessages;
 
   $: requireAccessControl = erc1155.isAccessControlRequired(opts);
 </script>
@@ -26,10 +29,10 @@
   </label>
   <label class="labeled-input">
     <span class="flex justify-between pr-2">
-      URI
+      Base URI
       <HelpTooltip>Location of the metadata. Clients will replace any instance of {"{id}"} in this string with the tokenId.</HelpTooltip>
     </span>
-    <input bind:value={opts.uri} placeholder="https://...">
+    <input bind:value={opts.baseUri} placeholder="https://...">
   </label>
 </section>
 
@@ -47,37 +50,31 @@
     <label class:checked={opts.burnable}>
       <input type="checkbox" bind:checked={opts.burnable}>
       Burnable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc1155#ERC1155Burnable">
-        Token holders will be able to destroy their tokens.
-      </HelpTooltip>
-    </label>
-    <label class:checked={opts.supply}>
-      <input type="checkbox" bind:checked={opts.supply}>
-      Supply Tracking
       <HelpTooltip>
-        Keeps track of total supply of tokens.
+        Token holders will be able to destroy their tokens.
       </HelpTooltip>
     </label>
     <label class:checked={opts.pausable}>
       <input type="checkbox" bind:checked={opts.pausable}>
       Pausable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/utils#Pausable">
-        Privileged accounts will be able to pause the functionality marked as <code>whenNotPaused</code>.
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/security#pausable">
+        Privileged accounts will be able to pause the functionality marked with <code>self.pausable.assert_not_paused()</code>.
         Useful for emergency response.
       </HelpTooltip>
     </label>
     <label class:checked={opts.updatableUri}>
       <input type="checkbox" bind:checked={opts.updatableUri}>
       Updatable URI
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc1155#ERC1155-_setURI-string-">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/api/erc1155#ERC1155Component-set_base_uri">
         Privileged accounts will be able to set a new URI for all token types. Clients will replace any instance of {"{id}"} in the URI with the tokenId.
       </HelpTooltip>
     </label>
+    <UpgradeabilityField bind:upgradeable={opts.upgradeable} />
   </div>
 </section>
 
-<AccessControlSection bind:access={opts.access} required={requireAccessControl} />
+<RoyaltyInfoSection bind:opts={opts.royaltyInfo} errors={errors} />
 
-<UpgradeabilitySection bind:upgradeable={opts.upgradeable} />
+<AccessControlSection bind:access={opts.access} required={requireAccessControl} />
 
 <InfoSection bind:info={opts.info} />
