+++ packages/ui/src/lang/cairo/ERC20Controls.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import HelpTooltip from '../HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
   import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo';
   import { premintPattern, erc20, infoDefaults } from '@openzeppelin/wizard-cairo';
@@ -7,7 +7,7 @@
   import AccessControlSection from './AccessControlSection.svelte';
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import InfoSection from './InfoSection.svelte';
-  import { error } from '../error-tooltip';
+  import { error } from '../../error-tooltip';
 
   export const opts: Required<KindedOptions['ERC20']> = {
     kind: 'ERC20',
