+++ packages/ui/src/lang/cairo/ERC721Controls.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import HelpTooltip from '../HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
   import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo';
   import { erc721, infoDefaults } from '@openzeppelin/wizard-cairo';
@@ -8,7 +8,7 @@
   import UpgradeabilityField from './UpgradeabilityField.svelte';
   import RoyaltyInfoSection from './RoyaltyInfoSection.svelte';
   import InfoSection from './InfoSection.svelte';
-  import { error } from '../error-tooltip';
+  import { error } from '../../error-tooltip';
 
   export const opts: Required<KindedOptions['ERC721']> = {
     kind: 'ERC721',
