+++ packages/ui/src/lang/cairo/InfoSection.svelte
