+++ packages/ui/src/lang/cairo/inject-hyperlinks.ts
