+++ packages/ui/src/lang/solidity/App.svelte
@@ -10,30 +10,30 @@
     import RealWorldAssetControls from './RealWorldAssetControls.svelte';
     import GovernorControls from './GovernorControls.svelte';
     import CustomControls from './CustomControls.svelte';
-    import CopyIcon from './icons/CopyIcon.svelte';
-    import CheckIcon from './icons/CheckIcon.svelte';
-    import RemixIcon from './icons/RemixIcon.svelte';
-    import DownloadIcon from './icons/DownloadIcon.svelte';
-    import ZipIcon from './icons/ZipIcon.svelte';
-    import FileIcon from './icons/FileIcon.svelte';
-    import ArrowsLeft from './icons/ArrowsLeft.svelte';
-    import ArrowsRight from './icons/ArrowsRight.svelte';
-    import Dropdown from './Dropdown.svelte';
-    import OverflowMenu from './OverflowMenu.svelte';
-    import Tooltip from './Tooltip.svelte';
+    import CopyIcon from '../../icons/CopyIcon.svelte';
+    import CheckIcon from '../../icons/CheckIcon.svelte';
+    import RemixIcon from '../../icons/RemixIcon.svelte';
+    import DownloadIcon from '../../icons/DownloadIcon.svelte';
+    import ZipIcon from '../../icons/ZipIcon.svelte';
+    import FileIcon from '../../icons/FileIcon.svelte';
+    import ArrowsLeft from '../../icons/ArrowsLeft.svelte';
+    import ArrowsRight from '../../icons/ArrowsRight.svelte';
+    import Dropdown from '../../Dropdown.svelte';
+    import OverflowMenu from '../../OverflowMenu.svelte';
+    import Tooltip from '../../Tooltip.svelte';
     import Wiz from './Wiz.svelte';
     import DefenderDeployModal from './DefenderDeployModal.svelte';
 
     import type { KindedOptions, Kind, Contract, OptionsErrorMessages } from '@openzeppelin/wizard';
     import { ContractBuilder, buildGeneric, printContract, sanitizeKind, OptionsError } from '@openzeppelin/wizard';
     import { getImports } from '@openzeppelin/wizard/get-imports';
-    import { postConfig } from './post-config';
+    import { postConfig } from '../../post-config';
     import { remixURL } from './remix';
 
     import { saveAs } from 'file-saver';
-    import { injectHyperlinks } from './utils/inject-hyperlinks';
+    import { injectHyperlinks } from './inject-hyperlinks';
     import { InitialOptions } from './initial-options';
-    import { postMessageToIframe } from './post-message';
+    import { postMessageToIframe } from '../../post-message';
 
     const dispatch = createEventDispatcher();
