+++ packages/ui/src/lang/solidity/CustomControls.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import HelpTooltip from './HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
   import type { KindedOptions } from '@openzeppelin/wizard';
   import { custom, infoDefaults } from '@openzeppelin/wizard';
