+++ packages/ui/src/lang/solidity/ERC1155Controls.svelte
@@ -1,12 +1,11 @@
 <script lang="ts">
-  import HelpTooltip from '../HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
-  import type { KindedOptions, OptionsErrorMessages } from '@openzeppelin/wizard-cairo';
-  import { erc1155, infoDefaults } from '@openzeppelin/wizard-cairo';
+  import type { KindedOptions } from '@openzeppelin/wizard';
+  import { erc1155, infoDefaults } from '@openzeppelin/wizard';
 
   import AccessControlSection from './AccessControlSection.svelte';
-  import UpgradeabilityField from './UpgradeabilityField.svelte';
-  import RoyaltyInfoSection from './RoyaltyInfoSection.svelte';
+  import UpgradeabilitySection from './UpgradeabilitySection.svelte';
   import InfoSection from './InfoSection.svelte';
 
   export let opts: Required<KindedOptions['ERC1155']> = {
@@ -14,8 +13,6 @@
     ...erc1155.defaults,
     info: { ...infoDefaults }, // create new object since Info is nested
   };
-
-  export let errors: undefined | OptionsErrorMessages;
 
   $: requireAccessControl = erc1155.isAccessControlRequired(opts);
 </script>
@@ -29,10 +26,10 @@
   </label>
   <label class="labeled-input">
     <span class="flex justify-between pr-2">
-      Base URI
+      URI
       <HelpTooltip>Location of the metadata. Clients will replace any instance of {"{id}"} in this string with the tokenId.</HelpTooltip>
     </span>
-    <input bind:value={opts.baseUri} placeholder="https://...">
+    <input bind:value={opts.uri} placeholder="https://...">
   </label>
 </section>
 
@@ -50,31 +47,37 @@
     <label class:checked={opts.burnable}>
       <input type="checkbox" bind:checked={opts.burnable}>
       Burnable
-      <HelpTooltip>
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc1155#ERC1155Burnable">
         Token holders will be able to destroy their tokens.
+      </HelpTooltip>
+    </label>
+    <label class:checked={opts.supply}>
+      <input type="checkbox" bind:checked={opts.supply}>
+      Supply Tracking
+      <HelpTooltip>
+        Keeps track of total supply of tokens.
       </HelpTooltip>
     </label>
     <label class:checked={opts.pausable}>
       <input type="checkbox" bind:checked={opts.pausable}>
       Pausable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/security#pausable">
-        Privileged accounts will be able to pause the functionality marked with <code>self.pausable.assert_not_paused()</code>.
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/utils#Pausable">
+        Privileged accounts will be able to pause the functionality marked as <code>whenNotPaused</code>.
         Useful for emergency response.
       </HelpTooltip>
     </label>
     <label class:checked={opts.updatableUri}>
       <input type="checkbox" bind:checked={opts.updatableUri}>
       Updatable URI
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts-cairo/api/erc1155#ERC1155Component-set_base_uri">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc1155#ERC1155-_setURI-string-">
         Privileged accounts will be able to set a new URI for all token types. Clients will replace any instance of {"{id}"} in the URI with the tokenId.
       </HelpTooltip>
     </label>
-    <UpgradeabilityField bind:upgradeable={opts.upgradeable} />
   </div>
 </section>
 
-<RoyaltyInfoSection bind:opts={opts.royaltyInfo} errors={errors} />
-
 <AccessControlSection bind:access={opts.access} required={requireAccessControl} />
+
+<UpgradeabilitySection bind:upgradeable={opts.upgradeable} />
 
 <InfoSection bind:info={opts.info} />
