+++ packages/ui/src/lang/solidity/ERC721Controls.svelte
@@ -1,46 +1,59 @@
 <script lang="ts">
-  import HelpTooltip from './HelpTooltip.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
   import type { KindedOptions } from '@openzeppelin/wizard';
-  import { erc20, premintPattern, infoDefaults } from '@openzeppelin/wizard';
-
+  import { erc721, infoDefaults } from '@openzeppelin/wizard';
+  
   import AccessControlSection from './AccessControlSection.svelte';
   import UpgradeabilitySection from './UpgradeabilitySection.svelte';
   import InfoSection from './InfoSection.svelte';
-  import ToggleRadio from './inputs/ToggleRadio.svelte';
-
-  export let opts: Required<KindedOptions['ERC20']> = {
-    kind: 'ERC20',
-    ...erc20.defaults,
-    premint: '', // default to empty premint in UI instead of 0
+  import ToggleRadio from '../../inputs/ToggleRadio.svelte';
+  
+  export let opts: Required<KindedOptions['ERC721']> = {
+    kind: 'ERC721',
+    ...erc721.defaults,
     info: { ...infoDefaults }, // create new object since Info is nested
   };
 
-  $: requireAccessControl = erc20.isAccessControlRequired(opts);
+  let wasMintable = opts.mintable;
+  let wasIncremental = opts.incremental;
+
+  $: {
+    if (wasMintable && !opts.mintable) {
+      opts.incremental = false;
+    }
+
+    if (opts.incremental && !wasIncremental) {
+      opts.mintable = true;
+    }
+
+    wasMintable = opts.mintable;
+    wasIncremental = opts.incremental;
+  }
+
+  $: requireAccessControl = erc721.isAccessControlRequired(opts);
 </script>
 
 <section class="controls-section">
   <h1>Settings</h1>
 
-    <div class="grid grid-cols-[2fr,1fr] gap-2">
-      <label class="labeled-input">
-        <span>Name</span>
-        <input bind:value={opts.name}>
-      </label>
-
-      <label class="labeled-input">
-        <span>Symbol</span>
-        <input bind:value={opts.symbol}>
-      </label>
-    </div>
-
+  <div class="grid grid-cols-[2fr,1fr] gap-2">
     <label class="labeled-input">
-      <span class="flex justify-between pr-2">
-        Premint
-        <HelpTooltip>Create an initial amount of tokens for the deployer.</HelpTooltip>
-      </span>
-      <input bind:value={opts.premint} placeholder="0" pattern={premintPattern.source}>
+      <span>Name</span>
+      <input bind:value={opts.name}>
     </label>
+    <label class="labeled-input">
+      <span>Symbol</span>
+      <input bind:value={opts.symbol}>
+    </label>
+  </div>
+  <label class="labeled-input">
+    <span class="flex justify-between pr-2">
+      Base URI
+      <HelpTooltip>Will be concatenated with token IDs to generate the token URIs.</HelpTooltip>
+    </span>
+    <input bind:value={opts.baseUri} placeholder="https://...">
+  </label>
 </section>
 
 <section class="controls-section">
@@ -51,18 +64,23 @@
       <input type="checkbox" bind:checked={opts.mintable}>
       Mintable
       <HelpTooltip>
-        Privileged accounts will be able to create more supply.
+        Privileged accounts will be able to emit new tokens.
       </HelpTooltip>
     </label>
-
+    <label class:checked={opts.incremental} class="subcontrol">
+      <input type="checkbox" bind:checked={opts.incremental}>
+      Auto Increment Ids
+      <HelpTooltip>
+        New tokens will be automatically assigned an incremental id.
+      </HelpTooltip>
+    </label>
     <label class:checked={opts.burnable}>
       <input type="checkbox" bind:checked={opts.burnable}>
       Burnable
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc20#ERC20Burnable">
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc721#ERC721Burnable">
         Token holders will be able to destroy their tokens.
       </HelpTooltip>
     </label>
-
     <label class:checked={opts.pausable}>
       <input type="checkbox" bind:checked={opts.pausable}>
       Pausable
@@ -71,20 +89,18 @@
         Useful for emergency response.
       </HelpTooltip>
     </label>
-
-    <label class:checked={opts.permit || opts.votes}>
-      <input type="checkbox" bind:checked={opts.permit}>
-      Permit
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc20#ERC20Permit">
-        Without paying gas, token holders will be able to allow third parties to transfer from their account.
+    <label class:checked={opts.enumerable}>
+      <input type="checkbox" bind:checked={opts.enumerable}>
+      Enumerable
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc721#ERC721Enumerable">
+        Allows on-chain enumeration of all tokens or those owned by an account. Increases gas cost of transfers.
       </HelpTooltip>
     </label>
-
-    <label class:checked={opts.flashmint}>
-      <input type="checkbox" bind:checked={opts.flashmint}>
-      Flash Minting
-      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc20#ERC20FlashMint">
-        Built-in flash loans. Lend tokens without requiring collateral as long as they're returned in the same transaction.
+    <label class:checked={opts.uriStorage}>
+      <input type="checkbox" bind:checked={opts.uriStorage}>
+      URI Storage
+      <HelpTooltip link="https://docs.openzeppelin.com/contracts/api/token/erc721#ERC721URIStorage">
+        Allows updating token URIs for individual token IDs.
       </HelpTooltip>
     </label>
   </div>
@@ -98,8 +114,8 @@
       <span class="ml-1">
         <ToggleRadio bind:value={opts.votes} defaultValue="blocknumber" />
       </span>
-      <HelpTooltip align="right" link="https://docs.openzeppelin.com/contracts/api/token/erc20#ERC20Votes">
-        Keeps track of historical balances for voting in on-chain governance, with a way to delegate one's voting power to a trusted account.
+      <HelpTooltip align="right" link="https://docs.openzeppelin.com/contracts/api/token/erc721#ERC721Votes">
+        Keeps track of individual units for voting in on-chain governance, with a way to delegate one's voting power to a trusted account.
       </HelpTooltip>
     </label>
   </h1>
