+++ packages/ui/src/lang/solidity/UpgradeabilitySection.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import type { Upgradeable } from '@openzeppelin/wizard';
 
-  import ToggleRadio from './inputs/ToggleRadio.svelte';
-  import HelpTooltip from './HelpTooltip.svelte';
+  import ToggleRadio from '../../inputs/ToggleRadio.svelte';
+  import HelpTooltip from '../../HelpTooltip.svelte';
 
   export let upgradeable: Upgradeable;
 </script>
