+++ packages/ui/src/lang/solidity/inject-hyperlinks.ts
